Consolidated all tab mapping / preparation to TabManager and cannibalized ActorSheetRuntimeManager.
Updated all callers.

As a side effect, item tabs now require a tab ID in order to be registered successfully.
